+++ src/fable-library-rust/src/Guid.rs
@@ -1,8 +1,9 @@
 #[cfg(feature = "guid")]
 pub mod Guid_ {
-    use crate::Native_::compare;
+    use crate::NativeArray_::{new_array, Array};
+    use crate::Native_::{compare, MutCell};
     use crate::String_::{string, toString};
-    use uuid::{Uuid};
+    use uuid::Uuid;
 
     #[derive(Clone, Copy, Debug, Default, PartialEq, PartialOrd)]
     pub struct Guid(Uuid);
@@ -27,10 +28,28 @@ pub mod Guid_ {
         Guid(Uuid::new_v4())
     }
 
-    pub fn parse(s: string) -> Guid {
-        match Uuid::parse_str(s.as_ref()) {
-            Ok(res) => Guid(res),
-            Err(x) => panic!("{}", x)
+    pub fn new_from_array(a: Array<u8>) -> Guid {
+        Guid(Uuid::from_slice_le(a.as_slice()).unwrap())
+    }
+
+    pub fn tryParse(s: string, res: &MutCell<Guid>) -> bool {
+        match Uuid::parse_str(s.trim()) {
+            Ok(uuid) => {
+                res.set(Guid(uuid));
+                true
+            }
+            Err(e) => false,
         }
+    }
+
+    pub fn parse(s: string) -> Guid {
+        match Uuid::parse_str(s.trim()) {
+            Ok(uuid) => Guid(uuid),
+            Err(e) => panic!("{}", e),
+        }
+    }
+
+    pub fn toByteArray(x: Guid) -> Array<u8> {
+        new_array(&x.0.to_bytes_le())
     }
 }